Changes in step file is reflected in feature file
This commit fixes the issue mentioned in issue#159, which
seems to have stopped working in 0.0.18. If you update a step file
(e.g.) changing an expression, the corresponding step in the feature
file will be marked as missing glue code. If you add a step for an
existing step, the warning regarding missing glue code in the feature
file will be removed.

The problem was that when cucumber.eclipse.steps.jdt.stepsBuilder ran
there was no available listeners that be triggered.

Getting it to work without cyclic dependencies required a lot of moving
around files.

Added is also the possibility to only search for step implementations
in the scope of the package where the feature file is located. Default
settings will have the same behaviour as before (e.g. it is backwards
compatible).